Funding task changes

+++ src/Dfe.ManageSchoolImprovement.Application/SupportProject/Commands/FundingHistory/EditFundingHistory.cs
@@ -10,7 +10,7 @@ public class EditFundingHistory
             FundingHistoryId Id,
             SupportProjectId SupportProjectId,
             string FundingType,
-            double FundingAmount,
+            decimal FundingAmount,
             string FinancialYear,
             int FundingRounds,
             string Comments
